perf(env): remove duplicate /api and Env class
