+++ libs/renderer/src/components/block-settings/custom/ChildBlockSettings.tsx
@@ -76,12 +76,12 @@ export const ChildBlockSettings = (props: ChildBlockSettingsProps) => {
         return <>Loading...</>;
     }
 
-    /**
-     * Initialize insight for app building
-     */
-    Env.update({
-        MODULE: process.env.MODULE || "",
-    });
+    // /**
+    //  * Initialize insight for app building
+    //  */
+    // Env.update({
+    //     MODULE: process.env.MODULE || "",
+    // });
 
     return (
         <Stack>
