--- packages/client/src/api/index.ts
@@ -1,2 +1,0 @@
-export * from './pixel';
-export * from './file';
