+++ packages/client/src/components/common/File/FileEditor.tsx
@@ -8,17 +8,17 @@ import {
     forwardRef,
     useImperativeHandle,
 } from 'react';
-import { styled, useNotification } from '@semoss/ui';
-
 import { OnMount } from '@monaco-editor/react';
 import parserBabel from 'prettier/parser-babel';
 import parserCss from 'prettier/parser-postcss';
 import parserHtml from 'prettier/parser-html';
 import prettier from 'prettier';
-const Editor = lazy(() => import('@monaco-editor/react'));
 
-import { runPixel } from '@/api';
+import { styled, useNotification } from '@semoss/ui';
 import { LoadingScreen } from '@/components/ui';
+import { runPixelTwo } from '@semoss/sdk';
+
+const Editor = lazy(() => import('@monaco-editor/react'));
 
 const IS_PRODUCTION = process.env.NODE_ENV == 'production';
 
@@ -203,7 +203,7 @@ export const FileEditor = forwardRef<FileEditorRefDef, FileEditorProps>(
                     throw new Error('Error missing pixel to get file');
                 }
 
-                const response = await runPixel<[string]>(pixel, insightId);
+                const response = await runPixelTwo<[string]>(pixel, insightId);
 
                 // set the content
                 const content = response.pixelReturn[0].output;
@@ -303,7 +303,7 @@ export const FileEditor = forwardRef<FileEditorRefDef, FileEditorProps>(
                     throw new Error('Error missing pixel to get file');
                 }
 
-                const { errors } = await runPixel(pixel, insightId);
+                const { errors } = await runPixelTwo(pixel, insightId);
 
                 // bubble up the errors
                 for (const e of errors) {
@@ -340,7 +340,7 @@ export const FileEditor = forwardRef<FileEditorRefDef, FileEditorProps>(
                     throw new Error('No Agent Model Engine');
                 }
 
-                const response = await runPixel(
+                const response = await runPixelTwo(
                     `LLM(engine = "${agentModelEngine}", command = "${prompt}", paramValues = [ {} ] );`,
                     insightId,
                 );
