+++ packages/client/src/pages/app/AppDetailPage.tsx
@@ -1,6 +1,18 @@
 import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useForm } from 'react-hook-form';
+import { Link } from 'react-router-dom';
+import {
+    Add,
+    Edit,
+    HdrAuto,
+    MoreVert,
+    Share,
+    EditLocation,
+    RemoveRedEyeRounded,
+    SimCardDownload,
+} from '@mui/icons-material';
+
 import {
     Breadcrumbs,
     Button,
@@ -14,6 +26,12 @@ import {
     Stack,
     CircularProgress,
 } from '@semoss/ui';
+import { Env } from '@semoss/sdk';
+
+import { useRootStore } from '@/hooks';
+import { formatPermission, toTitleCase } from '@/utility';
+import { ShareOverlay } from '@/components/ui';
+import { SettingsContext } from '@/contexts';
 import {
     MembersTable,
     PendingMembersTable,
@@ -35,23 +53,6 @@ import {
     DetailsForm,
     AppDetailsRef,
 } from '@/components/app';
-import { ShareOverlay } from '@/components/ui';
-import { SettingsContext } from '@/contexts';
-import { Env } from '@/env';
-import { useRootStore } from '@/hooks';
-import { formatPermission, toTitleCase } from '@/utility';
-import {
-    Add,
-    Edit,
-    HdrAuto,
-    MoreVert,
-    Share,
-    EditLocation,
-    RemoveRedEyeRounded,
-    SimCardDownload,
-} from '@mui/icons-material';
-
-import { Link } from 'react-router-dom';
 
 const OuterContainer = styled('div')(({ theme }) => ({
     backgroundColor: theme.palette.background.paper,
